chore: lazy loading of the map component
