+++ src/components/GeographyDrawer/index.tsx
@@ -1,23 +1,14 @@
-import { useState } from "react";
+import { Suspense, lazy, useState } from "react";
 import { iconClose, iconTarget } from "~/util/icons";
 import { ActionIcon, Drawer, Group } from "@mantine/core";
 import { Icon } from "~/components/Icon";
 import { Spacer } from "~/components/Spacer";
 import { ModalTitle } from "~/components/ModalTitle";
 import { DrawerResizer } from "~/components/DrawerResizer";
-import { MapContainer, Marker, TileLayer } from 'react-leaflet';
 import { GeometryPoint } from "surrealdb.js";
-import { LatLngTuple, Icon as LeafletIcon } from "leaflet";
-import 'leaflet/dist/leaflet.css';
-import markerIconPng from "leaflet/dist/images/marker-icon.png";
+import { LoadingContainer } from "../LoadingContainer";
 
-const markerIcon = new LeafletIcon({
-	iconUrl: markerIconPng,
-	iconSize: [25, 41],
-	iconAnchor: [12, 41],
-	popupAnchor: [1, -34],
-	shadowSize: [41, 41],
-});
+const GeographyMap = lazy(() => import("../GeographyMap"));
 
 export interface InspectorDrawerProps {
 	opened: boolean;
@@ -26,9 +17,6 @@ export interface InspectorDrawerProps {
 }
 
 export const GeographyDrawer = ({ opened, data, onClose }: InspectorDrawerProps) => {
-	const [long, lat] = data.point;
-	const point = [lat.toNumber(), long.toNumber()] as LatLngTuple;
-
 	const [width, setWidth] = useState(650);
 
 	return (
@@ -70,26 +58,11 @@ export const GeographyDrawer = ({ opened, data, onClose }: InspectorDrawerProps)
 				</Group>
 			</Group>
 
-			<section style={{
-				display: "flex",
-				justifyContent: "center",
-				alignItems: "center",
-				width: "100%",
-				height: "100%",
-			}}>
-				<MapContainer
-					zoom={13}
-					center={point}
-					style={{ width: "100%", height: "inherit" }}
-					scrollWheelZoom={false}
-					attributionControl={false}
-				>
-					<TileLayer
-						url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-					/>
-					<Marker icon={markerIcon} position={point} />
-				</MapContainer>
-			</section>
+			<Suspense
+				fallback={<LoadingContainer visible />}
+			>
+				<GeographyMap data={data} />
+			</Suspense>
 		</Drawer>
 	);
 };
